requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.3.0
tensorflow==2.13.0
reportlab==4.2.2
pyttsx3==2.90
opencv-python-headless==4.8.1.78
